Backend: refactored Idea structure, added interfaces for each service, fixed bugs Problems: reference chain from User[post]

+++ src/backend/java/org/dneversky/idea/repository/IdeaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IdeaRepository extends JpaRepository<Idea, Integer> {
+public interface IdeaRepository extends JpaRepository<Idea, Long> {
 
 }
